requirements.txt
ag2==0.2.0  # Autonomous Group 2 framework
openai==1.3.0  # For LLM capabilities
python-dotenv==1.0.0  # For environment variable management
aiohttp==3.9.1  # For async HTTP requests
pydantic==2.5.2  # For data validation
fastapi==0.105.0  # For API endpoints (if needed)
uvicorn==0.24.0  # ASGI server for FastAPI
sqlalchemy==2.0.23  # For database operations
alembic==1.13.0  # For database migrations
pytest==7.4.3  # For testing
pytest-asyncio==0.21.1  # For async tests
black==23.11.0  # For code formatting
mypy==1.7.1  # For type checking
python-jose==3.3.0  # For JWT handling
passlib==1.7.4  # For password hashing
bcrypt==4.1.1  # For password hashing
redis==5.0.1  # For caching and session management 